Crate swipc_gen

SwIPC Code Generator

This crate is responsible for generating client code from a SwIPC interface definition.

The client code will use sunrise_libuser's IPC API.

Modules

gen_rust_code
Code generation implementation

itemmod
Mod parser

Functions

generate_ipc
Generate a module containing all the functions in the given IPC file.