Function swipc_gen::gen_rust_code::get_type

fn get_type(
    output: bool, 
    ty: &Alias, 
    is_server: bool, 
    out_lifetime: &str
) -> Result<String, Error>

Get the Rust type of an Alias. If output is true, then the type should be suitable for a return type (or an output IPC buffer argument). If output is false, then the type should be suitable for an input argument.